(refactored) initializing store
