+++ src/routes/search/+page.svelte
@@ -1,51 +1,9 @@
 <script>
-	import { onMount } from 'svelte';
-	import { config } from '$lib/config';
-	import { db, search, fillResults } from '$lib/db';
+	import { db, search, fillResults, createFilterOptions, createIndex } from '$lib/db';
 	import DropdownFilter from '$lib/components/DropdownFilter.svelte';
 	import Results from '$lib/components/Results.svelte';
 	import Search from '$lib/icons/Search.svelte';
 	import ResetFilters from '$lib/components/ResetFilters.svelte';
-	import pkg from 'flexsearch';
-	const { Document } = pkg;
-
-	const index = new Document({ ...config.index });
-	const filterKeys = $db.filterKeys;
-
-	async function loadFilterOptions() {
-		let filters = {};
-		filterKeys.forEach((k) => {
-			filters = {
-				...filters,
-				[k]: new Set(
-					Object.values(index.store)
-						.map((e) => e[k])
-						.filter((e) => e !== undefined)
-				)
-			};
-		});
-		return filters;
-	}
-
-	async function loadIndex() {
-		const res = await fetch('/api/search-index');
-		const keys = await res.json();
-		for (const key in keys) {
-			await index.import(key, keys[key]);
-		}
-		const filters = await loadFilterOptions();
-		db.update((db) => {
-			return { ...db, filters };
-		});
-		db.update((db) => {
-			return { ...db, index };
-		});
-		fillResults();
-	}
-
-	onMount(() => {
-		loadIndex();
-	});
 </script>
 
 <div class="flex w-3/4 flex-col">
