+++ svelte.config.js
@@ -4,7 +4,16 @@ export default {
 	kit: {
 		adapter: adapter({
 			// see below for options that can be set here
-		})
+		}),
+		csp: {
+			directives: {
+				'script-src': ['self']
+			},
+			reportOnly: {
+				'script-src': ['self'],
+				'report-uri': ['/']
+			}
+		}
 	}
 };
